+++ shared/src/commonMain/kotlin/com/murzify/bambuddyspool/feature/tagmutation/TagMutationFeatureBoundary.kt
@@ -1,3 +1,0 @@
-package com.murzify.bambuddyspool.feature.tagmutation
-
-internal object TagMutationFeatureBoundary